+++ src/interfaces/ILagrangeService.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.12;
 
 interface ILagrangeService {
     function register(
-        uint256 chainID,
+        uint32 chainID,
         bytes memory _blsPubKey,
         uint32 serveUntilBlock
     ) external;
